quiz_project/requirements.txt
asgiref==3.8.1
coverage==7.12.0
diff-match-patch==20241021
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tablib==3.8.0
